PP-6361 Split charge_ft_tests.js some more
Split out charge_status_ft_tests.js,
card_details_send_data_ft_tests.js and
card_details_errors_ft_tests.js from charge_ft_tests.js.

No changes have been made to the tests: they’ve just been moved to
different files.

with @nimalank7
